chore: move matrix to common
